move stopwatch to be an island and fix product description
