+++ components/product/ProductCard.tsx
@@ -1,13 +1,14 @@
 import Image from "deco-sites/std/components/Image.tsx";
 import Icon from "$store/components/ui/Icon.tsx";
-import Stopwatch from "deco-sites/otica-isabela/components/product/Stopwatch.tsx";
-import ToExperimentButton from "deco-sites/otica-isabela/components/product/ToExperimentButton.tsx";
+import Stopwatch from "deco-sites/otica-isabela/islands/Stopwatch.tsx";
+import ToExperimentButton from "deco-sites/otica-isabela/islands/ToExperimentButton.tsx";
 import type { Product } from "deco-sites/std/commerce/types.ts";
 import { formatPrice } from "$store/sdk/format.ts";
 import { mapProductToAnalyticsItem } from "deco-sites/std/commerce/utils/productToAnalyticsItem.ts";
 import { SendEventOnClick } from "$store/sdk/analytics.tsx";
 import { getAvailableColors } from "deco-sites/otica-isabela/sdk/getVariantColors.ts";
 import { getDevice } from "deco-sites/otica-isabela/sdk/getDevice.ts";
+import { getDescriptions } from "deco-sites/otica-isabela/sdk/getDescriptions.ts";
 
 export interface Layout {
   basics?: {
@@ -70,48 +71,20 @@ function ProductCard({
 
   const id = `product-card-${productID}`;
 
-  const targetNames = [
-    "Altura da Lente",
-    "Largura da Lente",
-    "Largura da Ponte",
-    "Hastes",
-    "Frente Total",
-    "Aro",
-  ] as const;
-
-  const nameMapping = {
-    "Altura da Lente": "Altura",
-    "Largura da Lente": "Largura",
-    "Largura da Ponte": "Ponte",
-    "Frente Total": "Frente",
-  } as const;
-
   const [front] = images ?? [];
 
   const { highPrice: listPrice, lowPrice: price } = offers ?? {};
 
   const discount = Math.ceil(
-    (((listPrice ?? 0) - (price ?? 0)) / (listPrice ?? 0)) * 100
+    (((listPrice ?? 0) - (price ?? 0)) / (listPrice ?? 0)) * 100,
   );
 
-  const description = additionalProperty
-    ?.filter((property) =>
-      targetNames.some((targetName) => property?.name?.includes(targetName))
-    )
-    ?.map((property) => {
-      const mappedName =
-        nameMapping[property.name as keyof typeof nameMapping] || property.name;
-      return { ...property, name: mappedName };
-    });
-
+  const description = getDescriptions(additionalProperty!);
   const availableColors = getAvailableColors(product);
   const device = getDevice();
   const experimenterImage = additionalProperty?.find(
-    (prop) => prop.propertyID === "experimentador"
+    (prop) => prop.propertyID === "experimentador",
   )?.value;
-
-  //ToDo: remove this line and add a real date from loader
-  const now = new Date();
 
   return (
     <div
@@ -137,13 +110,7 @@ function ProductCard({
       />
 
       {/* Stopwatch */}
-      {isStopwatchEnabled && (
-        <Stopwatch
-          targetDate={
-            new Date(now.getFullYear(), now.getMonth(), now.getDate() + 10)
-          }
-        />
-      )}
+      {isStopwatchEnabled && <Stopwatch targetDate={new Date()} />}
 
       <figure class="relative" style={{ aspectRatio: `${306} / ${170}` }}>
         {/* Product Images */}
@@ -180,9 +147,9 @@ function ProductCard({
           <p class="text-sm font-normal text-base-200 line-clamp-3 min-h-[42px] mb-2 ">
             {description?.map(
               (property, index) =>
-                `${property?.name}: ${property?.value}mm ${
+                `${property?.value}: ${property?.name}mm ${
                   index < description.length - 1 ? "/" : ""
-                } `
+                } `,
             )}
           </p>
         </div>
@@ -194,10 +161,9 @@ function ProductCard({
               <a href={url} aria-label={name} title={name}>
                 <div
                   style={{
-                    background:
-                      unitCodes.length > 1
-                        ? `linear-gradient(${unitCodes.join(", ")})`
-                        : `${unitCodes[0]}`,
+                    background: unitCodes.length > 1
+                      ? `linear-gradient(${unitCodes.join(", ")})`
+                      : `${unitCodes[0]}`,
                   }}
                   class="mask mask-circle h-5 w-5 bg-secondary mx-2"
                 />
