+++ components/ui/NewModal.tsx
@@ -1,7 +1,7 @@
 import { useId } from "$store/sdk/useId.ts";
 import { useSignal } from "@preact/signals";
 import { ComponentChildren } from "preact";
-import { useEffect, useState } from "preact/hooks";
+import { useEffect } from "preact/hooks";
 
 interface Props {
   onClose?: () => void;
