+++ islands/Stopwatch.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/product/Stopwatch.tsx";
